% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{print_recipe}
\alias{print_recipe}
\title{Print a recipe}
\usage{
print_recipe(target, source, command = NULL)
}
\arguments{
\item{target}{a target filename.}

\item{source}{a vector of source filenames. NULL for empty dependencies.}

\item{command}{a vector of command lines. NULL for empty lines. Default: NULL.}
}
\description{
Print a recipe.
}
